+++ BLL/PropertyClasses/Transaction/OpeningStockProperty.cs
@@ -10,7 +10,7 @@ namespace BLL.PropertyClasses.Transaction
         public Int64 item_id { get; set; }
         public Int64 color_id { get; set; }
         public Int64 size_id { get; set; }
-        public Int64 opening_pcs { get; set; }
+        public decimal opening_pcs { get; set; }
         public decimal opening_rate { get; set; }
         #endregion
     }
